+++ src/Badge.js
@@ -1,0 +1,20 @@
+import sys from 'system-components'
+
+export const Badge = sys({
+  fontSize: 0,
+  px: 2,
+  py: 1,
+  mx: 1,
+  color: 'white',
+  bg: 'blue',
+  fontWeight: 'bold',
+  borderRadius: 2
+}, {
+  WebkitFontSmoothing: 'antialiased',
+  display: 'inline-block',
+  verticalAlign: 'middle'
+})
+
+Badge.displayName = 'Badge'
+
+export default Badge
