+++ src/Border.js
@@ -1,0 +1,10 @@
+import sys from 'system-components'
+
+export const Border = sys({
+  borderWidth: 1,
+  borderColor: 'gray'
+}, 'space', 'color')
+
+Border.displayName = 'Border'
+
+export default Border
