+++ src/ButtonTransparent.js
@@ -1,0 +1,27 @@
+import styled from 'styled-components'
+import { theme } from 'styled-system'
+import Button from './Button'
+import { darken } from './utils'
+
+export const ButtonTransparent = styled(Button)([], props => ({
+  '&:hover': {
+    color: theme('colors.' + props.color)(props),
+    backgroundColor: 'transparent'
+  },
+  '&:focus': {
+    boxShadow: `inset 0 0 0 2px, 0 0 0 2px`
+  },
+  '&:active': {
+    backgroundColor: 'transparent',
+    boxShadow: `inset 0 0 0 2px, inset 0 0 8px ${darken(1/4)}`
+  }
+}))
+
+ButtonTransparent.displayName = 'ButtonTransparent'
+
+ButtonTransparent.defaultProps = {
+  color: 'inherit',
+  bg: 'transparent'
+}
+
+export default ButtonTransparent
