+++ src/Card.js
@@ -1,0 +1,14 @@
+import sys from 'system-components'
+
+export const Card = sys({
+  p: 2,
+  bg: 'white',
+  borderRadius: 2,
+  boxShadow: 2
+}, {
+  overflow: 'hidden'
+})
+
+Card.displayName = 'Card'
+
+export default Card
