+++ src/Checkbox.js
@@ -1,0 +1,12 @@
+import sys from 'system-components'
+
+export const Checkbox = sys({
+  is: 'input',
+  type: 'checkbox',
+  m: 0,
+  mr: 2,
+})
+
+Checkbox.displayName = 'Checkbox'
+
+export default Checkbox
