+++ src/Label.js
@@ -1,0 +1,14 @@
+import sys from 'system-components'
+
+export const Label = sys({
+  is: 'label',
+  fontSize: 1,
+  mb: 1,
+  align: 'center'
+}, {
+  display: 'flex'
+}, 'alignItems')
+
+Label.displayName = 'Label'
+
+export default Label
