+++ src/Provider.js
@@ -1,0 +1,26 @@
+import React from 'react'
+import Root from './Root'
+import theme from './theme'
+
+export class Provider extends React.Component {
+  render () {
+    const {
+      theme,
+      ...props
+    } = this.props
+
+    return (
+      <ThemeProvider theme={theme}>
+        <Root {...props} />
+      </ThemeProvider>
+    )
+  }
+}
+
+Provider.defaultProps = {
+  theme
+}
+
+Provider.displayName = 'Rebass.Provider'
+
+export default Provider
