+++ src/Radio.js
@@ -1,0 +1,12 @@
+import sys from 'system-components'
+
+export const Radio = sys({
+  is: 'input',
+  type: 'radio',
+  m: 0,
+  mr: 2
+})
+
+Radio.displayName = 'Radio'
+
+export default Radio
