+++ src/Samp.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Code from './Code'
+
+export const Samp = props =>
+  <Code
+    is='samp'
+    {...props}
+  />
+
+Samp.displayName = 'Samp'
+
+export default Samp
